+++ src/PrestaShopBundle/Controller/Admin/Sell/Catalog/AttachmentController.php
@@ -31,7 +31,7 @@ use PrestaShop\PrestaShop\Core\Domain\Attachment\Command\DeleteAttachmentCommand
 use PrestaShop\PrestaShop\Core\Domain\Attachment\Exception\AttachmentException;
 use PrestaShop\PrestaShop\Core\Domain\Attachment\Exception\DeleteAttachmentException;
 use PrestaShop\PrestaShop\Core\Domain\Attachment\Query\AttachmentPath;
-use PrestaShop\PrestaShop\Core\Domain\Attachment\QueryResult\DownloadableAttachment;
+use PrestaShop\PrestaShop\Core\Domain\Attachment\QueryResult\Attachment;
 use PrestaShop\PrestaShop\Core\Search\Filters\AttachmentFilters;
 use PrestaShopBundle\Controller\Admin\FrameworkBundleAdminController;
 use PrestaShopBundle\Security\Annotation\AdminSecurity;
@@ -119,10 +119,10 @@ class AttachmentController extends FrameworkBundleAdminController
     public function viewAction($attachmentId)
     {
         try {
-            /** @var DownloadableAttachment $downloadableAttachment */
-            $downloadableAttachment = $this->getCommandBus()->handle(new AttachmentPath((int) $attachmentId));
+            /** @var Attachment $attachment */
+            $attachment = $this->getCommandBus()->handle(new AttachmentPath((int) $attachmentId));
 
-            return $this->file($downloadableAttachment->getPath(), $downloadableAttachment->getName());
+            return $this->file($attachment->getPath(), $attachment->getName());
         } catch (AttachmentException $e) {
             $this->addFlash('error', $this->getErrorMessageForException($e, $this->getErrorMessages()));
         }
