+++ src/PrestaShopBundle/Resources/config/services/adapter/attachment.yml
@@ -12,7 +12,7 @@ services:
     tags:
       - { name: 'tactician.handler', command: 'PrestaShop\PrestaShop\Core\Domain\Attachment\Command\BulkDeleteAttachmentsCommand' }
 
-  prestashop.adapter.contact.query_handler.get_attachment_path:
-    class: 'PrestaShop\PrestaShop\Adapter\Attachment\QueryHandler\AttachmentPathProvider'
+  prestashop.adapter.contact.query_handler.get_attachment_handler:
+    class: 'PrestaShop\PrestaShop\Adapter\Attachment\QueryHandler\GetAttachmentPathHandler'
     tags:
       - { name: 'tactician.handler', command: 'PrestaShop\PrestaShop\Core\Domain\Attachment\Query\AttachmentPath' }
